requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
elevenlabs==0.2.24
openai>=1.6.1
tweepy==4.12.1
python-socketio==5.7.2
requests==2.28.2
numpy==1.24.3
langchain==0.1.0
langchain-openai==0.0.2
